refactor: shared commands type
