+++ src/redux/thunks/previewReferencedHelmChart.ts
@@ -7,9 +7,9 @@ import {extractObjectsFromYaml} from '@redux/services/manifest-utils';
 import {interpolateTemplate} from '@redux/services/templates';
 import {createUnsavedResource} from '@redux/services/unsavedResource';
 
-import {CommandOptions, runCommandInMainThread} from '@utils/commands';
+import {runCommandInMainThread} from '@utils/commands';
 
-import {AppDispatch, K8sResource} from '@monokle-desktop/shared/models';
+import {AppDispatch, CommandOptions, K8sResource} from '@monokle-desktop/shared/models';
 
 const fsWriteFilePromise = promisify(fs.writeFile);
 const fsReadFilePromise = promisify(fs.readFile);
