+++ src/utils/kubeclient.ts
@@ -6,9 +6,9 @@ import {v4 as uuid} from 'uuid';
 
 import {getMainProcessEnv} from '@utils/env';
 
-import {ClusterAccess, KubePermissions} from '@monokle-desktop/shared/models';
+import {ClusterAccess, CommandOptions, CommandResult, KubePermissions} from '@monokle-desktop/shared/models';
 
-import {CommandOptions, CommandResult, runCommandInMainThread} from './commands/execute';
+import {runCommandInMainThread} from './commands/execute';
 import {isRendererThread} from './thread';
 
 export function createKubeClient(path: string, context?: string) {
